requirements.txt
pandas
polars
sqlalchemy
psycopg2-binary
requests